Update legacy.sh to build dual MBR/EFI media in all cases
This also keeps the "data" and dual code partitions and the semantics for online updates (e.g. updatep1/p2) by putting the EFI partition in slot 3 and using slot 4 as a "BSD" partition that then is divided into the CFG and DATA areas.

Additionally change the partitioning code to allocate on 1m boundaries (and work around a "funny" in gpart that can round down, which doesn't work when you want to "dd" in the created code partition) and make sure we put the correct MBR boot code in for CSM mode.  The result boots on a pcEngines device that has no clue what EFI is and I've used for years and also on a pair of EFI-only devices I have that can't boot "legacy" media at all.  In both cases a USB "stick" (if the only storage device on the USB ports) comes up as da0 which works great.

One fly in the ointment in the updatep1/p2 scripts (which are separate commits) is that I've found no reliable way to determine exactly what order the EFI loader enumerated the drives. This is serious because on EFI you set rootdev to determine which of the two code partitions boot and if the enumerated disk is wrong it doesn't boot at all leaving you dead at the EFI console.  If the ONLY media is the inserted stick it works fine but if not (e.g. there's an nVME or mSATA disk in the box without an OS on it, or with a different OS) you have potential trouble and I've not found a way from the running system (e.g. with sysctl or the EFI tools) to get the correct, "as the BIOS saw it when it booted", enumeration.